Only use relative links for jaxson.vandoorn.ca for local development
This resolves a bug for some links which quickly show a 404 page
prior to going to the page.  This issue also resulted in the back
button from these pages not working as the page would not reload.

This is due to Gatsby attempting to catch any relative links
even if they may not be part of the Gatsby site.